pyproject.toml
[project]
name = "udo-api"
version = "0.2.0"
description = "Unternehmensdaten API - Geodaten Service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.122.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "greenlet>=3.0.0",
    "pymssql>=2.3.0",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.10.1",
    "bcrypt>=5.0.0",
    "alembic>=1.18.1",
    "openpyxl>=3.1.5",
    "phonenumbers>=8.13",
    "pandas>=3.0.0",
    "xlrd>=2.0.2",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]
